more refactoring of actions and sequences
