+++ web-client/src/getConstants.js
@@ -32,6 +32,7 @@ import {
   EVENT_CODES_REQUIRING_SIGNATURE,
   FILING_TYPES,
   INITIAL_DOCUMENT_TYPES,
+  INITIAL_DOCUMENT_TYPES_MAP,
   IRS_SYSTEM_SECTION,
   MAX_FILE_SIZE_BYTES,
   MAX_FILE_SIZE_MB,
@@ -117,6 +118,7 @@ export const getConstants = () => ({
   EXTERNAL_USER_DASHBOARD_TABS,
   FILING_TYPES,
   INITIAL_DOCUMENT_TYPES,
+  INITIAL_DOCUMENT_TYPES_MAP,
   INTERNAL_CATEGORY_MAP: DOCUMENT_INTERNAL_CATEGORIES_MAP,
   IRS_SYSTEM_SECTION,
   MAX_FILE_SIZE_BYTES,
